Add 03_HR1
Tanja sent over a new dataset with 4/12 files to be transcribed.
No excel this time, segmentation is already done.
This time I implemented an in-memory segmentation.
Slow as heck, but it works. I'm thinking of keeping it, it's not as messy
and it works also if the segments would have a duplicated name (as was the case in this
corpus.)